+++ soxspipe/commonutils/__init__.py
@@ -19,4 +19,5 @@ from .uncompress import uncompress
 from .horne_extraction import horne_extraction
 from .response_function import response_function
 from .flux_calibration import flux_calibration
+from .reducer import reducer
 # from .set_of_files import set_of_files