+++ soxspipe/recipes/soxs_order_centres.py
@@ -65,7 +65,7 @@ class soxs_order_centres(_base_recipe_):
         super(soxs_order_centres, self).__init__(
             log=log, settings=settings, inputFrames=inputFrames, overwrite=overwrite, recipeName="soxs-order-centre")
         self.log = log
-        log.debug("instansiating a new 'soxs_order_centres' object")
+        log.debug("instantiating a new 'soxs_order_centres' object")
         self.settings = settings
         self.inputFrames = inputFrames
         self.verbose = verbose