+++ soxspipe/recipes/soxs_spatial_solution.py
@@ -57,7 +57,7 @@ class soxs_spatial_solution(_base_recipe_):
         super(soxs_spatial_solution, self).__init__(
             log=log, settings=settings, inputFrames=inputFrames, overwrite=overwrite, recipeName="soxs-spatial-solution")
         self.log = log
-        log.debug("instansiating a new 'soxs_spatial_solution' object")
+        log.debug("instantiating a new 'soxs_spatial_solution' object")
         self.settings = settings
         self.inputFrames = inputFrames
         self.verbose = verbose